+++ app/src/main/java/ru/krivocraft/kbmp/core/track/TracksAdapter.java
@@ -17,11 +17,11 @@ import com.devs.vectorchildfinder.VectorDrawableCompat;
 
 import java.util.Collections;
 
+import ru.krivocraft.kbmp.R;
 import ru.krivocraft.kbmp.contexts.MainActivity;
+import ru.krivocraft.kbmp.contexts.TrackEditorActivity;
 import ru.krivocraft.kbmp.core.ColorManager;
 import ru.krivocraft.kbmp.core.ItemTouchHelperAdapter;
-import ru.krivocraft.kbmp.R;
-import ru.krivocraft.kbmp.contexts.TrackEditorActivity;
 import ru.krivocraft.kbmp.core.playback.MediaService;
 import ru.krivocraft.kbmp.core.storage.TracksStorageManager;
 import ru.krivocraft.kbmp.tasks.LoadArtTask;
