+++ app/src/main/java/ru/krivocraft/kbmp/fragments/SmallPlayerFragment.java
@@ -29,13 +29,13 @@ import java.util.Objects;
 import java.util.Timer;
 import java.util.TimerTask;
 
-import ru.krivocraft.kbmp.core.ColorManager;
 import ru.krivocraft.kbmp.R;
-import ru.krivocraft.kbmp.core.playback.MediaService;
-import ru.krivocraft.kbmp.core.utils.TimeUtils;
 import ru.krivocraft.kbmp.contexts.PlayerActivity;
+import ru.krivocraft.kbmp.core.ColorManager;
+import ru.krivocraft.kbmp.core.playback.MediaService;
 import ru.krivocraft.kbmp.core.storage.TracksStorageManager;
 import ru.krivocraft.kbmp.core.utils.BitmapUtils;
+import ru.krivocraft.kbmp.core.utils.TimeUtils;
 
 public class SmallPlayerFragment extends BaseFragment {
