+++ app/src/main/java/ru/krivocraft/kbmp/sqlite/DBConnection.java
@@ -4,16 +4,23 @@ import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.database.sqlite.SQLiteOpenHelper;
+
+import androidx.annotation.Nullable;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import ru.krivocraft.kbmp.core.ColorManager;
+import ru.krivocraft.kbmp.core.storage.TrackListsStorageManager;
 import ru.krivocraft.kbmp.core.track.Track;
 import ru.krivocraft.kbmp.core.track.TrackList;
 import ru.krivocraft.kbmp.core.track.TrackReference;
 
 public class DBConnection {
+    private static final String TRACKS = "tracks";
+    private static final String TRACK_LISTS = "track_lists";
+    private static final String ALL_TRACKS = TrackList.createIdentifier(TrackListsStorageManager.STORAGE_TRACKS_DISPLAY_NAME);
     private SQLiteDatabase database;
 
     public DBConnection(Context context) {
@@ -31,7 +38,7 @@ public class DBConnection {
         values.put("selected", track.isSelected() ? 1 : 0);
         values.put("artist", track.getArtist());
         values.put("path", track.getPath());
-        database.insert(TableNames.TRACKS, null, values);
+        database.insert(TRACKS, null, values);
     }
 
     public void updateTrack(Track track) {
@@ -44,21 +51,21 @@ public class DBConnection {
         values.put("playing", track.isPlaying() ? 1 : 0);
         values.put("liked", track.isLiked() ? 1 : 0);
         values.put("selected", track.isSelected() ? 1 : 0);
-        database.update(TableNames.TRACKS, values, "id = ?", new String[]{String.valueOf(track.getIdentifier())});
+        database.update(TRACKS, values, "id = ?", new String[]{String.valueOf(track.getIdentifier())});
     }
 
     public void removeTrackList(TrackList trackList) {
-        database.delete(TableNames.TRACK_LISTS, "id = ?", new String[]{trackList.getIdentifier()});
+        database.delete(TRACK_LISTS, "id = ?", new String[]{trackList.getIdentifier()});
         database.execSQL("drop table if exists '" + trackList.getIdentifier() + "'");
     }
 
     public void removeTrack(Track track) {
-        database.delete(TableNames.TRACKS, "id = ?", new String[]{String.valueOf(track.getIdentifier())});
+        database.delete(TRACKS, "id = ?", new String[]{String.valueOf(track.getIdentifier())});
     }
 
     public Track getTrack(TrackReference trackReference) {
         Track track = new Track(0, "", "", "", ColorManager.GREEN);
-        Cursor cursor = database.query(TableNames.TRACKS, null, "id = ?", new String[]{trackReference.toString()}, null, null, null);
+        Cursor cursor = database.query(TRACKS, null, "id = ?", new String[]{trackReference.toString()}, null, null, null);
         if (cursor.moveToFirst()) {
 
             long duration = cursor.getLong(cursor.getColumnIndex("duration"));
@@ -78,7 +85,7 @@ public class DBConnection {
 
     public List<String> getTrackListNames() {
         List<String> identifiers = new ArrayList<>();
-        Cursor cursor = database.query(TableNames.TRACK_LISTS, new String[]{"name"}, null, null, null, null, null);
+        Cursor cursor = database.query(TRACK_LISTS, new String[]{"name"}, null, null, null, null, null);
         if (cursor.moveToFirst()) {
             do {
                 identifiers.add(cursor.getString(cursor.getColumnIndex("name")));
@@ -101,7 +108,7 @@ public class DBConnection {
     public void updateTrackListData(TrackList trackList) {
         ContentValues values = new ContentValues();
         values.put("name", trackList.getDisplayName());
-        database.update(TableNames.TRACK_LISTS, values, "id = ?", new String[]{trackList.getIdentifier()});
+        database.update(TRACK_LISTS, values, "id = ?", new String[]{trackList.getIdentifier()});
     }
 
     public void updateTrackListContent(TrackList trackList) {
@@ -115,7 +122,7 @@ public class DBConnection {
 
     public List<Track> getTracksStorage() {
         List<Track> tracks = new ArrayList<>();
-        Cursor cursor = database.query(TableNames.TRACKS, null, null, null, null, null, null);
+        Cursor cursor = database.query(TRACKS, null, null, null, null, null, null);
         if (cursor.moveToFirst()) {
             int titleIndex = cursor.getColumnIndex("title");
             int artistIndex = cursor.getColumnIndex("artist");
@@ -145,7 +152,7 @@ public class DBConnection {
 
     public List<TrackList> getTrackLists() {
         List<TrackList> trackLists = new ArrayList<>();
-        Cursor cursor = database.query(TableNames.TRACK_LISTS, null, null, null, null, null, null);
+        Cursor cursor = database.query(TRACK_LISTS, null, null, null, null, null, null);
         if (cursor.moveToFirst()) {
             int idIndex = cursor.getColumnIndex("id");
             int nameIndex = cursor.getColumnIndex("name");
@@ -178,7 +185,7 @@ public class DBConnection {
             selection.append("type != " + TrackList.TRACK_LIST_BY_TAG);
         }
 
-        Cursor cursor = database.query(TableNames.TRACK_LISTS, null, selection.toString(), null, null, null, null);
+        Cursor cursor = database.query(TRACK_LISTS, null, selection.toString(), null, null, null, null);
         if (cursor.moveToFirst()) {
             int idIndex = cursor.getColumnIndex("id");
             int nameIndex = cursor.getColumnIndex("name");
@@ -199,7 +206,7 @@ public class DBConnection {
 
     public TrackList getTrackList(String identifier) {
         TrackList trackList = null;
-        Cursor cursor = database.query(TableNames.TRACK_LISTS, null, "id = ?", new String[]{identifier}, null, null, null);
+        Cursor cursor = database.query(TRACK_LISTS, null, "id = ?", new String[]{identifier}, null, null, null);
         if (cursor.moveToFirst()) {
             int idIndex = cursor.getColumnIndex("id");
             int nameIndex = cursor.getColumnIndex("name");
@@ -222,7 +229,7 @@ public class DBConnection {
             values.put("id", trackList.getIdentifier());
             values.put("name", trackList.getDisplayName());
             values.put("type", trackList.getType());
-            database.insert(TableNames.TRACK_LISTS, null, values);
+            database.insert(TRACK_LISTS, null, values);
         }
     }
 
@@ -262,4 +269,84 @@ public class DBConnection {
         return trackReferences;
     }
 
+    public static class DBHelper extends SQLiteOpenHelper {
+
+        private final ColorManager colorManager;
+
+        DBHelper(@Nullable Context context) {
+            super(context, "tracks", null, 2);
+            this.colorManager = new ColorManager(context);
+        }
+
+        @Override
+        public void onCreate(SQLiteDatabase db) {
+            db.execSQL("create table if not exists " + TRACKS + " ("
+                    + "id integer,"
+                    + "playing integer,"
+                    + "selected integer,"
+                    + "liked integer,"
+                    + "title text,"
+                    + "color integer,"
+                    + "artist text,"
+                    + "duration long,"
+                    + "path text);");
+            db.execSQL("create table if not exists " + TRACK_LISTS + " ("
+                    + "id text,"
+                    + "name text,"
+                    + "type integer);");
+            db.execSQL("create table if not exists " + ALL_TRACKS + " ("
+                    + "id integer primary key autoincrement,"
+                    + "reference integer);");
+
+            ContentValues values = new ContentValues();
+            values.put("id", TrackList.createIdentifier(TrackListsStorageManager.STORAGE_TRACKS_DISPLAY_NAME));
+            values.put("name", TrackListsStorageManager.STORAGE_TRACKS_DISPLAY_NAME);
+            values.put("type", TrackList.TRACK_LIST_CUSTOM);
+            db.insert(TRACK_LISTS, null, values);
+        }
+
+        @Override
+        public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+            try {
+                db.execSQL("SELECT color FROM " + TRACKS);
+            } catch (Exception e) {
+                db.execSQL("alter table " + TRACKS + " add column color integer default " + ColorManager.GREEN);
+                generateColors(db);
+            }
+        }
+
+        private void updateColor(SQLiteDatabase database, Track track) {
+            ContentValues values = new ContentValues();
+            values.put("color", colorManager.getRandomColor());
+            database.update(TRACKS, values, "id = ?", new String[]{String.valueOf(track.getIdentifier())});
+        }
+
+        private void generateColors(SQLiteDatabase db) {
+            Cursor cursor = db.query(TRACKS, null, null, null, null, null, null);
+            if (cursor.moveToFirst()) {
+                int titleIndex = cursor.getColumnIndex("title");
+                int artistIndex = cursor.getColumnIndex("artist");
+                int pathIndex = cursor.getColumnIndex("path");
+                int durationIndex = cursor.getColumnIndex("duration");
+                int colorIndex = cursor.getColumnIndex("color");
+                int likedIndex = cursor.getColumnIndex("liked");
+                int selectedIndex = cursor.getColumnIndex("selected");
+                int playingIndex = cursor.getColumnIndex("playing");
+                do {
+                    long duration = cursor.getLong(durationIndex);
+                    String artist = cursor.getString(artistIndex);
+                    String title = cursor.getString(titleIndex);
+                    String path = cursor.getString(pathIndex);
+                    boolean liked = cursor.getInt(likedIndex) == 1;
+                    boolean selected = cursor.getInt(selectedIndex) == 1;
+                    boolean playing = cursor.getInt(playingIndex) == 1;
+                    int color = cursor.getInt(colorIndex);
+
+                    Track track = new Track(duration, artist, title, path, liked, selected, playing, color);
+                    updateColor(db, track);
+                } while (cursor.moveToNext());
+            }
+            cursor.close();
+        }
+    }
 }
